Add TestUserClock
