+++ app/src/androidTest/java/org/simple/clinic/TestClinicApp.kt
@@ -31,9 +31,11 @@ import org.simple.clinic.storage.StorageModule
 import org.simple.clinic.sync.SyncScheduler
 import org.simple.clinic.user.LoggedInUserHttpInterceptor
 import org.simple.clinic.user.UserSession
+import org.simple.clinic.util.TestUserClock
 import org.simple.clinic.util.TestUtcClock
+import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.UtcClock
-import org.threeten.bp.Clock
+import org.threeten.bp.ZoneId
 import timber.log.Timber
 import java.util.UUID
 import javax.inject.Inject
@@ -69,6 +71,8 @@ class TestClinicApp : ClinicApp() {
     return DaggerTestAppComponent.builder()
         .appModule(object : AppModule(this) {
           override fun utcClock(): UtcClock = TestUtcClock()
+
+          override fun userClock(userTimeZone: ZoneId): UserClock = TestUserClock()
         })
         .storageModule(object : StorageModule(databaseName = "ignored", runDatabaseQueriesOnMainThread = true) {
           override fun sqliteOpenHelperFactory() = AppSqliteOpenHelperFactory(inMemory = true)
