+++ app/src/main/java/org/simple/clinic/di/AppModule.kt
@@ -16,6 +16,7 @@ import org.simple.clinic.storage.StorageModule
 import org.simple.clinic.summary.PatientSummaryModule
 import org.simple.clinic.sync.SyncModule
 import org.simple.clinic.sync.indicator.SyncIndicatorModule
+import org.simple.clinic.util.RealUserClock
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.UtcClock
 import org.threeten.bp.ZoneId
@@ -54,7 +55,7 @@ open class AppModule(private val appContext: Application) {
 
   @Provides
   @AppScope
-  fun userClock(userTimeZone: ZoneId) = UserClock(userTimeZone)
+  open fun userClock(userTimeZone: ZoneId): UserClock= RealUserClock(userTimeZone)
 
   @Provides
   @AppScope
