+++ app/src/main/java/org/simple/clinic/util/Clocks.kt
@@ -1,26 +1,29 @@
 package org.simple.clinic.util
 
 import org.threeten.bp.Clock
+import org.threeten.bp.Instant
 import org.threeten.bp.ZoneId
 
 open class UtcClock : Clock() {
 
   private val utcClock = Clock.systemUTC()
 
-  override fun withZone(zone: ZoneId) = utcClock.withZone(zone)!!
+  override fun withZone(zone: ZoneId): Clock = utcClock.withZone(zone)
 
-  override fun getZone() = utcClock.zone!!
+  override fun getZone(): ZoneId = utcClock.zone
 
-  override fun instant() = utcClock.instant()!!
+  override fun instant(): Instant = utcClock.instant()
 }
 
-open class UserClock(userTimeZone: ZoneId) : Clock() {
+abstract class UserClock: Clock()
+
+open class RealUserClock(userTimeZone: ZoneId) : UserClock() {
 
   private val userClock = Clock.system(userTimeZone)
 
-  override fun withZone(zone: ZoneId) = userClock.withZone(zone)!!
+  override fun withZone(zone: ZoneId): Clock = userClock.withZone(zone)
 
-  override fun getZone() = userClock.zone!!
+  override fun getZone(): ZoneId = userClock.zone
 
-  override fun instant() = userClock.instant()!!
+  override fun instant(): Instant = userClock.instant()
 }
